CLAIM-298: Addressed review comments.
